feat: Add CSS styles to complete the holy grail mockup task
This was an interesting one and it was quite hard in the beginning. It
was great looking at all the possibilities that could be done with
Grids. There was a point I'd wanted to use Flexbox for some of the
obvious things but as the project requested, I used Grid throughout.

This was a great test of my grid understanding.